requirements.txt
pandas
ipykernel
matplotlib
openpyxl
sentence_transformers
faiss-cpu
tf-keras
openai